Fix autostart Generation
Fixed autostart file generation and wrapped the deletion in an if to check if the file exists first.

+++ setupKiosk.sh
@@ -39,11 +39,13 @@ sudo apt-get -y install --no-install-recommends xserver-xorg x11-xserver-utils x
 
 #Configure Openbox autostart
 	#Remove default autostart file
-	sudo rm /etc/xdg/openbox/autostart
+	if [ -f /etc/xdg/openbox/autostart ]; then
+		sudo rm /etc/xdg/openbox/autostart
+	fi
 	
 	#Create a new autostart with the information we want
 
-	sudo cat <<EOT >>/etc/xdg/openbox/autostart
+	sudo sh -c 'cat > /etc/xdg/openbox/autostart' << EOF
 	#cat <<EOT >>/home/pi/test.txt
 #Disable any form of screen saver / screen blanking / power management
 xset s off
@@ -57,7 +59,7 @@ setxkbmap -option terminate:ctrl_alt_bksp
 sed -i 's/"exited_cleanly":false/"exited_cleanly":true' ~/.config/chromium/'Local State'
 sed -i 's/"exited_cleanly":false/"exited_cleanly":true; s/"exit_type":"[^"]\+"/"exit_type":"Normal"/' ~/.config/chromium/Default/Preferences
 chromium-browser --disable-infobars --kiosk '$KIOSKURL'
-EOT
+EOF
 
 #Create .bash_profile
 rm /home/pi/.bash_profile
